refactor(project): splits schema out into multiple files (#68)
