+++ lib/project/schema/project_go_gen.cue
@@ -1,0 +1,51 @@
+// Code generated by cue get go. DO NOT EDIT.
+
+//cue:generate cue get go github.com/input-output-hk/catalyst-forge/lib/project/schema
+
+package schema
+
+// Project contains the configuration for the project.
+#Project: {
+	// Name contains the name of the project.
+	name: string @go(Name)
+
+	// Container is the name that the container will be built as.
+	container: string @go(Container)
+
+	// CI contains the configuration for the CI system.
+	// +optional
+	ci?: #ProjectCI @go(CI)
+
+	// Deployment contains the configuration for the deployment of the project.
+	// +optional
+	deployment?: #Deployment @go(Deployment)
+}
+
+#ProjectCI: {
+	// Targets configures the individual targets that are run by the CI system.
+	// +optional
+	targets?: {[string]: #Target} @go(Targets,map[string]Target)
+}
+
+// Target contains the configuration for a single target.
+#Target: {
+	// Args contains the arguments to pass to the target.
+	// +optional
+	args?: {[string]: string} @go(Args,map[string]string)
+
+	// Platforms contains the platforms to run the target against.
+	// +optional
+	platforms?: [...string] @go(Platforms,[]string)
+
+	// Privileged determines if the target should run in privileged mode.
+	// +optional
+	privileged?: null | bool @go(Privileged,*bool)
+
+	// Retries contains the number of times to retry the target.
+	// +optional
+	retries?: null | int @go(Retries,*int)
+
+	// Secrets contains the secrets to pass to the target.
+	// +optional
+	secrets?: [...#Secret] @go(Secrets,[]Secret)
+}
