+++ lib/project/schema/providers_go_gen.cue
@@ -1,0 +1,84 @@
+// Code generated by cue get go. DO NOT EDIT.
+
+//cue:generate cue get go github.com/input-output-hk/catalyst-forge/lib/project/schema
+
+package schema
+
+// Providers contains the configuration for the providers being used by the CI system.
+#Providers: {
+	// AWS contains the configuration for the AWS provider.
+	// +optional
+	aws?: #ProviderAWS @go(AWS)
+
+	// Docker contains the configuration for the DockerHub provider.
+	// +optional
+	docker?: #ProviderDocker @go(Docker)
+
+	// Earthly contains the configuration for the Earthly Cloud provider.
+	// +optional
+	earthly?: #ProviderEarthly @go(Earthly)
+
+	// Git contains the configuration for the Git provider.
+	// +optional
+	git?: #ProviderGit @go(Git)
+
+	// Github contains the configuration for the Github provider.
+	// +optional
+	github?: #ProviderGithub @go(Github)
+}
+
+// ProviderAWS contains the configuration for the AWS provider.
+#ProviderAWS: {
+	// Role contains the role to assume.
+	role?: null | string @go(Role,*string)
+
+	// Region contains the region to use.
+	region?: null | string @go(Region,*string)
+
+	// Registry contains the ECR registry to use.
+	// +optional
+	registry?: null | string @go(Registry,*string)
+}
+
+// ProviderDocker contains the configuration for the DockerHub provider.
+#ProviderDocker: {
+	// Credentials contains the credentials to use for DockerHub
+	credentials: #Secret @go(Credentials)
+}
+
+// ProviderEarthly contains the configuration for the Earthly Cloud provider.
+#ProviderEarthly: {
+	// Credentials contains the credentials to use for Earthly Cloud
+	// +optional
+	credentials?: #Secret @go(Credentials)
+
+	// Org specifies the Earthly Cloud organization to use.
+	// +optional
+	org?: null | string @go(Org,*string)
+
+	// Satellite contains the satellite to use for caching.
+	// +optional
+	satellite?: null | string @go(Satellite,*string)
+
+	// The version of Earthly to use in CI.
+	// +optional
+	version?: null | string @go(Version,*string)
+}
+
+// ProviderGit contains the configuration for the Git provider.
+#ProviderGit: {
+	// Credentials contains the credentials to use for interacting with private repositories.
+	// +optional
+	credentials?: null | #Secret @go(Credentials,*Secret)
+}
+
+// ProviderGithub contains the configuration for the Github provider.
+#ProviderGithub: {
+	// Credentials contains the credentials to use for Github
+	//  +optional
+	credentials?: #Secret @go(Credentials)
+
+	// Registry contains the Github registry to use.
+	// +optional
+	registry?: null | string @go(Registry,*string)
+}
